+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   provided.
 - `UnicornEmulator` captures more detailed error information in single step
   mode.
+- `UnicornEmulator.write_memory()` now supports overlapping writes and no
+  longer requires addresses to be page aligned.
 
 ## [0.0.1] - 2024-02-26